feat(shared-libs): add messaging capability package
