+++ shared-libs/ECommerce.Shared.DeadLetter/ECommerce.Shared.DeadLetter.csproj
@@ -12,7 +12,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\ECommerce.Shared.Kernel\ECommerce.Shared.Kernel.csproj" />
-    <ProjectReference Include="..\ECommerce.Shared.EventBus\ECommerce.Shared.EventBus.csproj" />
+    <ProjectReference Include="..\ECommerce.Shared.Messaging\ECommerce.Shared.Messaging.csproj" />
     <ProjectReference Include="..\ECommerce.Shared.RabbitMq\ECommerce.Shared.RabbitMq.csproj" />
     <ProjectReference Include="..\ECommerce.Shared.AzureServiceBus\ECommerce.Shared.AzureServiceBus.csproj" />
     <ProjectReference Include="..\ECommerce.Shared.LayoutAnalyzer\ECommerce.Shared.LayoutAnalyzer.csproj"
