+++ shared-libs/ECommerce.Shared.Messaging/Abstractions/.placeholder
@@ -1,0 +1,1 @@
+Keep this directory so the Messaging package keeps the shared-libs Abstractions/Impl/Composition shape.
