+++ shared-libs/ECommerce.Shared.Messaging/Composition/MessagingStartupExtensions.cs
