+++ shared-libs/ECommerce.Shared.Messaging/ECommerce.Shared.Messaging.csproj
@@ -1,0 +1,22 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <ItemGroup>
+    <InternalsVisibleTo Include="ECommerce.Shared.DeadLetter" />
+    <InternalsVisibleTo Include="ECommerce.Shared.Messaging.Tests" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\ECommerce.Shared.Kernel\ECommerce.Shared.Kernel.csproj" />
+    <ProjectReference Include="..\ECommerce.Shared.EventBus\ECommerce.Shared.EventBus.csproj" />
+    <ProjectReference Include="..\ECommerce.Shared.RabbitMq\ECommerce.Shared.RabbitMq.csproj" />
+    <ProjectReference Include="..\ECommerce.Shared.AzureServiceBus\ECommerce.Shared.AzureServiceBus.csproj" />
+    <ProjectReference Include="..\ECommerce.Shared.LayoutAnalyzer\ECommerce.Shared.LayoutAnalyzer.csproj"
+                      OutputItemType="Analyzer"
+                      ReferenceOutputAssembly="false" />
+  </ItemGroup>
+
+</Project>
