+++ shared-libs/ECommerce.Shared.Messaging/Impl/MessagingProviderResolver.cs
