+++ shared-libs/ECommerce.Shared.slnx
@@ -4,6 +4,7 @@
   <Project Path="ECommerce.Shared.EventBus/ECommerce.Shared.EventBus.csproj" />
   <Project Path="ECommerce.Shared.RabbitMq/ECommerce.Shared.RabbitMq.csproj" />
   <Project Path="ECommerce.Shared.AzureServiceBus/ECommerce.Shared.AzureServiceBus.csproj" />
+  <Project Path="ECommerce.Shared.Messaging/ECommerce.Shared.Messaging.csproj" />
   <Project Path="ECommerce.Shared.DeadLetter/ECommerce.Shared.DeadLetter.csproj" />
   <Project Path="ECommerce.Shared.Platform/ECommerce.Shared.Platform.csproj" />
   <Project Path="ECommerce.Shared.Contracts/ECommerce.Shared.Contracts.csproj" />
@@ -13,6 +14,7 @@
   <Project Path="tests/ECommerce.Shared.EventBus.Tests/ECommerce.Shared.EventBus.Tests.csproj" />
   <Project Path="tests/ECommerce.Shared.RabbitMq.Tests/ECommerce.Shared.RabbitMq.Tests.csproj" />
   <Project Path="tests/ECommerce.Shared.AzureServiceBus.Tests/ECommerce.Shared.AzureServiceBus.Tests.csproj" />
+  <Project Path="tests/ECommerce.Shared.Messaging.Tests/ECommerce.Shared.Messaging.Tests.csproj" />
   <Project Path="tests/ECommerce.Shared.DeadLetter.Tests/ECommerce.Shared.DeadLetter.Tests.csproj" />
   <Project Path="tests/ECommerce.Shared.Platform.Tests/ECommerce.Shared.Platform.Tests.csproj" />
   <Project Path="tests/ECommerce.Shared.Contracts.Tests/ECommerce.Shared.Contracts.Tests.csproj" />
