+++ shared-libs/ECommerce.Shared/ECommerce.Shared.csproj
@@ -5,6 +5,7 @@
     <ProjectReference Include="..\ECommerce.Shared.EventBus\ECommerce.Shared.EventBus.csproj" />
     <ProjectReference Include="..\ECommerce.Shared.RabbitMq\ECommerce.Shared.RabbitMq.csproj" />
     <ProjectReference Include="..\ECommerce.Shared.AzureServiceBus\ECommerce.Shared.AzureServiceBus.csproj" />
+    <ProjectReference Include="..\ECommerce.Shared.Messaging\ECommerce.Shared.Messaging.csproj" />
     <ProjectReference Include="..\ECommerce.Shared.DeadLetter\ECommerce.Shared.DeadLetter.csproj" />
     <ProjectReference Include="..\ECommerce.Shared.Platform\ECommerce.Shared.Platform.csproj" />
     <ProjectReference Include="..\ECommerce.Shared.Contracts\ECommerce.Shared.Contracts.csproj" />
