+++ shared-libs/tests/ECommerce.Shared.Messaging.Tests/ECommerce.Shared.Messaging.Tests.csproj
@@ -9,11 +9,8 @@
 
   <ItemGroup>
     <PackageReference Include="coverlet.collector" Version="6.0.4" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="10.0.5" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.14.1" />
     <PackageReference Include="NetArchTest.Rules" Version="1.3.2" />
-    <PackageReference Include="NSubstitute" Version="5.3.0" />
-    <PackageReference Include="Testcontainers.RabbitMq" Version="4.11.0" />
     <PackageReference Include="xunit" Version="2.9.3" />
     <PackageReference Include="xunit.runner.visualstudio" Version="3.1.4" />
   </ItemGroup>
@@ -27,11 +24,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\ECommerce.Shared.RabbitMq\ECommerce.Shared.RabbitMq.csproj" />
-    <!-- RabbitMqDeadLetterIntegrationTests verifies the RabbitMq branch of the DLQ capture/replay
-         pipeline and saga command DLQ flow, which live in the DeadLetter + Contracts packages. -->
-    <ProjectReference Include="..\..\ECommerce.Shared.DeadLetter\ECommerce.Shared.DeadLetter.csproj" />
-    <ProjectReference Include="..\..\ECommerce.Shared.Contracts\ECommerce.Shared.Contracts.csproj" />
+    <ProjectReference Include="..\..\ECommerce.Shared.Messaging\ECommerce.Shared.Messaging.csproj" />
   </ItemGroup>
 
 </Project>
